fastapi-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
httpx==0.24.0
python-dotenv==1.0.0
requests==2.28.2
pandas==2.0.0
scikit-learn==1.2.2